requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-cors-headers
dj-database-url
djangorestframework-simplejwt